+++ server/src/test/java/org/elasticsearch/tasks/TaskManagerTests.java
@@ -20,16 +20,21 @@
 package org.elasticsearch.tasks;
 
 import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.action.admin.cluster.node.tasks.TransportTasksActionTests;
 import org.elasticsearch.common.lease.Releasable;
 import org.elasticsearch.common.lease.Releasables;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.common.util.concurrent.ConcurrentCollections;
 import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.threadpool.TestThreadPool;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.FakeTcpChannel;
 import org.elasticsearch.transport.TcpChannel;
 import org.elasticsearch.transport.TransportRequest;
+import org.elasticsearch.transport.TransportService;
+import org.junit.After;
+import org.junit.Before;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -45,6 +50,18 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.mock;
 
 public class TaskManagerTests extends ESTestCase {
+    private ThreadPool threadPool;
+
+    @Before
+    public void setupThreadPool() {
+        threadPool = new TestThreadPool(TransportTasksActionTests.class.getSimpleName());
+    }
+
+    @After
+    public void terminateThreadPool() {
+        terminate(threadPool);
+    }
+
     /**
      * Makes sure that tasks that attempt to store themselves on completion retry if
      * they don't succeed at first.
@@ -58,11 +75,14 @@ public class TaskManagerTests extends ESTestCase {
         assertEquals(600000L, total);
     }
 
-    public void testTrackingChannelTask() {
-        final TaskManager taskManager = new TaskManager(Settings.EMPTY, mock(ThreadPool.class), Set.of());
+    public void testTrackingChannelTask() throws Exception {
+        final TaskManager taskManager = new TaskManager(Settings.EMPTY, threadPool, Set.of());
         Set<CancellableTask> cancelledTasks = new HashSet<>();
-        taskManager.setOrphanedTasksOnChannelCloseListener(tasks -> {
-            for (CancellableTask task : tasks) {
+        taskManager.setTaskCancellationService(new TaskCancellationService(mock(TransportService.class)) {
+            @Override
+            void cancelTaskAndDescendants(CancellableTask task, String reason, boolean waitForCompletion, ActionListener<Void> listener) {
+                assertThat(reason, equalTo("channel was closed"));
+                assertFalse(waitForCompletion);
                 assertTrue("task [" + task + "] was cancelled already", cancelledTasks.add(task));
             }
         });
@@ -94,17 +114,20 @@ public class TaskManagerTests extends ESTestCase {
             } else {
                 expectedCancelledTasks.add(task);
             }
-            assertThat(cancelledTasks, equalTo(expectedCancelledTasks));
         }
-        assertThat(cancelledTasks, equalTo(expectedCancelledTasks));
+        assertBusy(() -> assertThat(cancelledTasks, equalTo(expectedCancelledTasks)));
+        for (FakeTcpChannel channel : channels) {
+            channel.close();
+        }
         assertThat(taskManager.numberOfChannelPendingTaskTrackers(), equalTo(0));
     }
 
     public void testTrackingTaskAndCloseChannelConcurrently() throws Exception {
-        final TaskManager taskManager = new TaskManager(Settings.EMPTY, mock(ThreadPool.class), Set.of());
+        final TaskManager taskManager = new TaskManager(Settings.EMPTY, threadPool, Set.of());
         Set<CancellableTask> cancelledTasks = ConcurrentCollections.newConcurrentSet();
-        taskManager.setOrphanedTasksOnChannelCloseListener(tasks -> {
-            for (CancellableTask task : tasks) {
+        taskManager.setTaskCancellationService(new TaskCancellationService(mock(TransportService.class)) {
+            @Override
+            void cancelTaskAndDescendants(CancellableTask task, String reason, boolean waitForCompletion, ActionListener<Void> listener) {
                 assertTrue("task [" + task + "] was cancelled already", cancelledTasks.add(task));
             }
         });
@@ -138,7 +161,7 @@ public class TaskManagerTests extends ESTestCase {
         for (Thread thread : threads) {
             thread.join();
         }
-        assertThat(cancelledTasks, equalTo(expectedCancelledTasks));
+        assertBusy(() -> assertThat(cancelledTasks, equalTo(expectedCancelledTasks)));
         assertThat(taskManager.numberOfChannelPendingTaskTrackers(), equalTo(0));
     }
